requirements.txt
aiobotocore
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aioitertools
aiosignal
alabaster
altair
annotated-types==0.7.0
anyio
appdirs==1.4.4
argon2-cffi
argon2-cffi-bindings
arrow
astroid
astropy
asttokens
async-lru
atomicwrites==1.4.0
attrs
Automat
autopep8
Babel
backports.functools-lru-cache
backports.tempfile
backports.weakref==1.0.post1
banks==2.1.1
bcrypt
beautifulsoup4==4.13.3
binaryornot
black
bleach
blinker
bokeh
boltons
botocore
Bottleneck
Brotli
cachetools
certifi
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
colorcet
comm
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cytoolz
dask
dataclasses-json==0.6.7
datashader
debugpy
decorator
defusedxml
Deprecated==1.2.18
deprecation==2.1.0
diff-match-patch
dill
dirtyjson==1.0.8
distributed
distro
docstring-to-markdown
docutils
entrypoints
et-xmlfile==1.1.0
executing
fastapi==0.115.12
fastjsonschema
filelock
filetype==1.2.0
flake8
Flask
fonttools==4.25.0
frozendict
frozenlist
fsspec
future
gensim
gitdb
GitPython
gmpy2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gotrue==2.12.0
greenlet
griffe==1.7.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
h5py
HeapDict
hf-xet==1.1.3
holoviews
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.5
hvplot
hyperframe==6.1.0
hyperlink
idna
imagecodecs
imageio
imagesize
imbalanced-learn
importlib-metadata
Pillow==9.5.0
google-ai-generativelanguage == 0.6.15
google-api-core == 2.25.0
google-api-python-client == 2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0 
grpcio-status==1.71.0
python-dotenv==1.1.0
fastapi
uvicorn
python-multipart  # needed for file uploads